+++ src/main/java/com/sbproject/piclocket/PicLocketApplication.java
@@ -1,6 +1,7 @@
 package com.sbproject.piclocket;
 
 import com.sbproject.piclocket.config.AwsProperties;
+import com.sbproject.piclocket.config.CorsProperties;
 import com.sbproject.piclocket.security.JwtProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -8,7 +9,7 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 
 @SpringBootApplication
 @EnableConfigurationProperties({
-        AwsProperties.class, JwtProperties.class
+        AwsProperties.class, JwtProperties.class, CorsProperties.class
 })
 public class PicLocketApplication {